requirement.txt
# cat requirements.txt
# update 30/10/2019 0h36

numpy==1.16.5
opencv-python==4.1.0.25
matplotlib==3.1.1
keras==2.3.0
tensorflow==2.0.0
